newsAggregatorSystem/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
requests==2.26.0
beautifulsoup4==4.9.3
lxml==4.6.3
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
scikit-learn==0.24.2
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
gunicorn==20.1.0